requirements.txt
bresenham==0.2.1
click==8.1.7
dataclasses-json==0.6.1
matplotlib==3.8.0
networkx==3.1
numpy==1.26.0
pathfinding==1.0.4
perlin-noise==1.12
pygame==2.5.2
python-statemachine==2.1.1
questionary==2.0.1
scikit-learn==1.3.1
scipy==1.11.3
sortedcontainers==2.4.0
tqdm==4.66.1
chex==0.1.83
dm-haiku==0.0.10
jax==0.4.18
jaxlib==0.4.18  # https://jax.readthedocs.io/en/latest/installation.html#nvidia-gpu
# pip install "jax[cuda12_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
lz4==4.3.2
optax==0.1.7
tensorboardX==2.6.2.2